feat: implimented sound length timers

+++ includes/apu.h
@@ -53,7 +53,7 @@ typedef struct {
     bool dac_enabled; // true iff (NRx2 & 0xF8) != 0
 
     uint8_t wave_ram[SAMPLES / 2]; // wave ram, each byte holds two samples
-    uint8_t length_timer;
+    uint16_t length_timer;
 
     uint16_t period_divider;
     uint8_t sample_index;
